+++ db.json
@@ -1,19 +1,19 @@
 {
   "todos": [
     {
-      "reply": "456",
+      "reply": "2352525555355",
+      "edit": false,
+      "id": 1
+    },
+    {
+      "reply": "1111111",
       "edit": false,
       "id": 2
     },
     {
-      "reply": "에네르기파!",
+      "reply": "야호야호",
       "edit": false,
       "id": 3
-    },
-    {
-      "reply": "315351515",
-      "edit": false,
-      "id": 4
     }
   ]
 }
