+++ src/App.js
@@ -1,30 +1,16 @@
 import './App.css';
 import { useSelector, useDispatch } from 'react-redux';
-import Reply from './Reply/Reply.jsx'
-import Thunk from './Reply/Thunk';
-import Prac from './Prac';
+import Reply from './Reply/Reply';
 
 
 function App() {
-
-const state = useSelector((state)=>state)
-
-const dispatch = useDispatch()
-
 
 
   return (
     <div className="App">
       <h1>Hello, react!</h1>
-      {/* <button onClick={ confirm }>Redux 확인 버튼</button> */}
 
-      {/* <Reply/>
-
-      <hr/>
-
-      <Thunk/> */}
-
-      <Prac/>
+      <Reply/>
 
     </div>
   );
