+++ src/Reply/Reply.module.css
@@ -1,0 +1,27 @@
+.containerReply {
+    border: 2px solid #D7D7D7;
+    width: 500px;
+    height: 600px;
+  
+    margin: 0 auto 0 auto;
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .replyFlex1 {
+    height: 500px;
+    padding: 10px;
+    
+    overflow: auto;
+  }
+  
+    .replyText {
+      padding: 10px;
+      border-bottom: 1px solid gray;
+    }
+  
+  .replyFlex2 {
+    height: 100px;
+    border-top: 2px solid #D7D7D7;
+    padding: 13px;
+  }
